+++ tests/phpunit/tests/REST_API/Stories_Controller.php
@@ -20,8 +20,7 @@ namespace Google\Web_Stories\Tests\REST_API;
 use Google\Web_Stories\Experiments;
 use Google\Web_Stories\Settings;
 use Google\Web_Stories\Story_Post_Type;
-use Google\Web_Stories\KSES;
-use Google\Web_Stories\Tests\KsesSetup;
+use Google\Web_Stories\Tests\Kses_Setup;
 use Spy_REST_Server;
 use WP_REST_Request;
 
@@ -29,7 +28,7 @@ use WP_REST_Request;
  * @coversDefaultClass \Google\Web_Stories\REST_API\Stories_Controller
  */
 class Stories_Controller extends \WP_Test_REST_TestCase {
-	use KsesSetup;
+	use Kses_Setup;
 
 	protected $server;
